requirements.txt
fastapi
uvicorn
librosa
numpy
scikit-learn
joblib
soundfile
python-multipart
python-dotenv
tqdm